Upgrade v0.9.9

+++ python/cog/server/worker.py
@@ -11,12 +11,21 @@ from typing import Any, Dict, Iterable, Optional, TextIO, Union
 from sentry_sdk import capture_exception
 
 from ..json import make_encodeable
-from ..predictor import (BasePredictor, get_predict, load_predictor_from_ref,
-                         run_setup)
-from .eventtypes import (Done, Heartbeat, Log, PredictionInput,
-                         PredictionOutput, PredictionOutputType, Shutdown)
-from .exceptions import (CancelationException, FatalWorkerException,
-                         InvalidStateException)
+from ..predictor import BasePredictor, get_predict, load_predictor_from_ref, run_setup
+from .eventtypes import (
+    Done,
+    Heartbeat,
+    Log,
+    PredictionInput,
+    PredictionOutput,
+    PredictionOutputType,
+    Shutdown,
+)
+from .exceptions import (
+    CancelationException,
+    FatalWorkerException,
+    InvalidStateException,
+)
 from .helpers import StreamRedirector, WrappedStream
 
 _spawn = multiprocessing.get_context("spawn")
@@ -81,7 +90,11 @@ class Worker:
             self._child.join()
 
     def cancel(self) -> None:
-        if self._allow_cancel and self._child.is_alive():
+        if (
+            self._allow_cancel
+            and self._child.is_alive()
+            and self._child.pid is not None
+        ):
             os.kill(self._child.pid, signal.SIGUSR1)
             self._allow_cancel = False
